refactor: @Enumerated 추가
